+++ cpg-analysis/src/main/kotlin/de/fraunhofer/aisec/cpg/graph/Extensions.kt
@@ -1,0 +1,43 @@
+/*
+ * Copyright (c) 2021, Fraunhofer AISEC. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ *                    $$$$$$\  $$$$$$$\   $$$$$$\
+ *                   $$  __$$\ $$  __$$\ $$  __$$\
+ *                   $$ /  \__|$$ |  $$ |$$ /  \__|
+ *                   $$ |      $$$$$$$  |$$ |$$$$\
+ *                   $$ |      $$  ____/ $$ |\_$$ |
+ *                   $$ |  $$\ $$ |      $$ |  $$ |
+ *                   \$$$$$   |$$ |      \$$$$$   |
+ *                    \______/ \__|       \______/
+ *
+ */
+package de.fraunhofer.aisec.cpg.graph
+
+import de.fraunhofer.aisec.cpg.graph.declarations.Declaration
+import de.fraunhofer.aisec.cpg.graph.statements.expressions.ArrayCreationExpression
+import de.fraunhofer.aisec.cpg.graph.statements.expressions.Expression
+
+fun Expression.evaluate(): Any? {
+    return ValueEvaluator().evaluate(this)
+}
+
+fun Declaration.evaluate(): Any? {
+    return ValueEvaluator().evaluateDeclaration(this)
+}
+
+val ArrayCreationExpression.capacity: Int
+    get() {
+        return dimensions.first().evaluate() as Int
+    }
